% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album.R
\name{album-methods}
\alias{album-methods}
\alias{raw_album_search}
\alias{raw_album_get_info}
\title{Album Methods}
\usage{
raw_album_search(album, .start_page = 1, .n_pages = NULL)

raw_album_get_info(
  artist = NULL,
  album = NULL,
  mbid = NULL,
  autocorrect = TRUE,
  username = NULL,
  lang = NULL
)
}
\arguments{
\item{album}{Album name as a string.}

\item{.start_page}{The page to start pulling from, by default starts from the
1st page}

\item{.n_pages}{The total number of pages to pull. If no value is provided,
all pages are retrieved.}

\item{artist}{Artist name as a string.}

\item{mbid}{(Optional) MusicBrainz ID for the album as a string. Can be provided instead of artist and album names}

\item{autocorrect}{(Optional) Whether to correct the artist/album/track name names provided in the response. Defaults to TRUE if not provided.}

\item{lang}{(Optional) The biography language.}
}
\value{
\itemize{
\item \code{\link[=raw_album_search]{raw_album_search()}} returns a list of lastfmr objects
\item \code{\link[=raw_album_get_info]{raw_album_get_info()}} returns a single lastfmr object
}
}
\description{
The methods are used to retrieve information about albums:
\itemize{
\item \code{\link[=raw_album_get_info]{raw_album_get_info()}} - Search for an album by name.
\item \code{\link[=raw_album_search]{raw_album_search()}} - Retrieve information about an album exactly by artist and album name or MBID. If both are provided the Artist and Album name are used by default.
}
}
\section{Language Setting}{


Biographies for albums and artists are returned in English by
default. To get the biography in other languages, provide a \href{https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes}{ISO 639-1 country code}
via the \code{lang} argument of supporting functions.
}

\section{Name Autocorrection}{


Endpoints that take artist/track/album names try to auto-correct misspelled names
in the search query. This behaviour can be overriden by calling the function
with argument \code{autocorrect} set to \code{FALSE}.
}

