% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album.R
\name{album_getInfo}
\alias{album_getInfo}
\title{Get the metadata and tracklist for an album.}
\usage{
album_getInfo(artist, album, mbid = NA, autocorrect = NA,
  username = NA, lang = NA)
}
\arguments{
\item{artist}{The artist name.}

\item{album}{The album name.}

\item{mbid}{The musicbrainz id for the album.}

\item{autocorrect}{Transform misspelled artist names into correct artist names,
returning the correct version instead.
The corrected artist name will be returned in the response. [0|1]}

\item{username}{The username for the context of the request.
If supplied, the user's playcount for this album is included in the response.}

\item{lang}{The language to return the biography in,
expressed as an ISO 639 alpha-2 code.}
}
\value{
A list of the metadata and tracklist for an album.
}
\description{
Get the metadata and tracklist for an album on Last.fm using the album name or a musicbrainz id.
Implementation of last.fm's \emph{album.getInfo} API method
(\url{http://www.last.fm/api/show/album.getInfo})
}
\examples{
\dontrun{
album_getInfo("Father John Misty", "Fear Fun")
}
}
