% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album.R
\name{album_getTopTags}
\alias{album_getTopTags}
\title{Get the top tags for an album, ordered by popularity.}
\usage{
album_getTopTags(artist, album, mbid = NA, autocorrect = NA)
}
\arguments{
\item{artist}{The artist name.}

\item{album}{The album name.}

\item{mbid}{The musicbrainz id for the album.}

\item{autocorrect}{Transform misspelled artist names into correct artist names,
returning the correct version instead.
The corrected artist name will be returned in the response. [0|1]}
}
\value{
A list of the top tags for an album.
}
\description{
Get the top tags for an album on Last.fm, ordered by popularity.
Implementation of last.fm's \emph{album.getTopTags} API method
(\url{http://www.last.fm/api/show/album.getTopTags})
}
\examples{
album_getTopTags("Miles Davis", "Kind of Blue")
}
