% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album.R
\name{album_search}
\alias{album_search}
\title{Search for an album by name.}
\usage{
album_search(album, limit = NA, page = NA)
}
\arguments{
\item{album}{The album name.}

\item{limit}{The number of results to fetch per page. Defaults to 50.}

\item{page}{The page number you wish to scan to.}
}
\value{
A list of the searched albums.
}
\description{
Search for an album by name. Returns album matches sorted by relevance.
Implementation of last.fm's \emph{album.search} API method
(\url{http://www.last.fm/api/show/album.search})
}
\examples{
\dontrun{
album_search("Rhythm & Reason")
}
}
