% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album.R
\name{albumGetInfo}
\alias{albumGetInfo}
\title{Get the metadata and tracklist for an album}
\usage{
albumGetInfo(api_key, artist = NULL, album = NULL, mbid = NULL,
  autocorrect = TRUE, user = NULL, lang = NULL)
}
\arguments{
\item{api_key}{(apiAccount; Required) : A Last.fm API key.}

\item{artist}{(character; Required, unless mbid) : The artist name.}

\item{album}{(character; Required, unless mbid) : The album name.}

\item{mbid}{(character; Optional) : The musicbrainz id for the album}

\item{autocorrect}{(logical; Optional) : Transform misspelled artist
names into correct artist names, returning the correct version instead.
The corrected artist name will be returned in the response.}

\item{user}{(apiUser; Optional) : The username for the context of the
request. If supplied, the user's playcount for this album is included
in the response.}

\item{lang}{(Optional) : The language to return the biography in,
expressed as an ISO 639 alpha-2 code.}
}
\value{
A \code{\link[tibble]{tibble}} of results
}
\description{
Get the metadata and tracklist for an album on Last.fm using the album
name or a musicbrainz id.
}
\references{
See \url{https://www.last.fm/api/show/album.getInfo}
}
