% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{apiAccount}
\alias{apiAccount}
\title{Make an \code{apiAccount} object.}
\usage{
apiAccount(api_key)
}
\arguments{
\item{api_key}{A string representing a Last.fm API key.}
}
\value{
An \code{apiAccount}.
}
\description{
Make an \code{apiAccount} object.
}
