% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{apiUser}
\alias{apiUser}
\title{Make an \code{apiUser} object.}
\usage{
apiUser(user, api_key)
}
\arguments{
\item{user}{A string representing a user's account on Last.fm}

\item{api_key}{An object returned from \code{\link{apiAccount}}}
}
\value{
An \code{apiUser}.
}
\description{
Make an \code{apiUser} object.
}
