% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-artist.R
\name{artist-methods}
\alias{artist-methods}
\alias{raw_artist_search}
\alias{raw_artist_get_info}
\alias{raw_artist_get_similar}
\alias{raw_artist_get_tags}
\alias{raw_artist_get_top_albums}
\alias{raw_artist_get_top_tracks}
\title{Artist API Methods}
\usage{
raw_artist_search(artist, .start_page = 1, .n_pages = NULL)

raw_artist_get_info(
  artist = NULL,
  mbid = NULL,
  autocorrect = TRUE,
  username = NULL,
  lang = NULL
)

raw_artist_get_similar(artist = NULL, mbid = NULL, autocorrect = TRUE)

raw_artist_get_tags(
  artist = NULL,
  mbid = NULL,
  autocorrect = TRUE,
  username = NULL
)

raw_artist_get_top_albums(
  artist = NULL,
  mbid = NULL,
  autocorrect = TRUE,
  .start_page = 1,
  .n_pages = NULL
)

raw_artist_get_top_tracks(
  artist = NULL,
  mbid = NULL,
  autocorrect = TRUE,
  .start_page = 1,
  .n_pages = NULL
)
}
\arguments{
\item{artist}{Artist name as a string.}

\item{.start_page}{The page to start pulling from, by default starts from the
1st page}

\item{.n_pages}{The total number of pages to pull. If no value is provided,
all pages are retrieved.}

\item{mbid}{MusicBrainzID as a string, this is optional unless the artist is
not provided.}

\item{autocorrect}{(Optional) Whether to correct the artist/album/track name names provided in the response. Defaults to TRUE if not provided.}

\item{username}{(Optional) If provided, the playcount of the artist by the
user is included in the response}

\item{lang}{(Optional) The biography language.}
}
\value{
\itemize{
\item \code{\link[=raw_artist_search]{raw_artist_search()}}, \code{\link[=raw_artist_get_top_albums]{raw_artist_get_top_albums()}}, and \code{\link[=raw_artist_get_top_tracks]{raw_artist_get_top_tracks()}} return a list of lastfmr object.
\item \code{\link[=raw_artist_get_info]{raw_artist_get_info()}}, \code{\link[=raw_artist_get_similar]{raw_artist_get_similar()}}, and \code{\link[=raw_artist_get_tags]{raw_artist_get_tags()}} return a single lastfmr object.
}
}
\description{
These endpoints return information about artists. Identifying artists can be
done either using the artist name or a MusicBrainzID. If both \code{artist} and
\code{mbid} are provided, the artist name is used by default.
\itemize{
\item \code{\link[=raw_artist_search]{raw_artist_search()}} - Search for an artist by name.
\item \code{\link[=raw_artist_get_info]{raw_artist_get_info()}} - Retrieve information about an artist exactly by name or MBID.
\item \code{\link[=raw_artist_get_similar]{raw_artist_get_similar()}} - Search for artists which are similar to a given artist.
\item \code{\link[=raw_artist_get_tags]{raw_artist_get_tags()}} - Get the tags applied by an individual user to an artist on Last.fm.
\item \code{\link[=raw_artist_get_top_albums]{raw_artist_get_top_albums()}} and \code{\link[=raw_artist_get_top_tracks]{raw_artist_get_top_tracks()}} - Get the top albums and tracks for an artist on Last.fm respectively.
}
}
\section{Language Setting}{


Biographies for albums and artists are returned in English by
default. To get the biography in other languages, provide a \href{https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes}{ISO 639-1 country code}
via the \code{lang} argument of supporting functions.
}

\section{Name Autocorrection}{


Endpoints that take artist/track/album names try to auto-correct misspelled names
in the search query. This behaviour can be overriden by calling the function
with argument \code{autocorrect} set to \code{FALSE}.
}

