% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist.R
\name{artist_getCorrection}
\alias{artist_getCorrection}
\title{Get corrections of an artist name.}
\usage{
artist_getCorrection(artist)
}
\arguments{
\item{artist}{The artist name.}
}
\value{
A list of corrections.
}
\description{
Use the last.fm corrections data to check whether the supplied artist has a
correction to a canonical artist.
Implementation of last.fm's \emph{artist.getCorrection} API method
(\url{http://www.last.fm/api/show/artist.getCorrection})
}
\examples{
\dontrun{
artist_getCorrection("guns and roses")
}
}
