% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist.R
\name{artist_getSimilar}
\alias{artist_getSimilar}
\title{Get all the artists similar to this artist.}
\usage{
artist_getSimilar(artist, mbid = NA, autocorrect = NA, limit = NA)
}
\arguments{
\item{artist}{The artist name.}

\item{mbid}{The musicbrainz id for the artist.}

\item{autocorrect}{Transform misspelled artist names into correct artist
names, returning the correct version instead. The corrected artist name will
be returned in the response. [0|1]}

\item{limit}{Limit the number of similar artists returned.}
}
\value{
A list of artists.
}
\description{
Get all the artists similar to this artist.
Implementation of last.fm's \emph{artist.getSimilar} API method
(\url{http://www.last.fm/api/show/artist.getSimilar})
}
\examples{
artist_getSimilar("Moses Sumney")
}
