% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist.R
\name{artist_getTags}
\alias{artist_getTags}
\title{Get the tags applied by an individual user to an artist.}
\usage{
artist_getTags(artist, user, mbid = NA, autocorrect = NA)
}
\arguments{
\item{artist}{The artist name.}

\item{user}{Username.}

\item{mbid}{The musicbrainz id for the artist.}

\item{autocorrect}{Transform misspelled artist names into correct artist
names, returning the correct version instead. The corrected artist name will
be returned in the response. [0|1]}
}
\value{
A list of tags.
}
\description{
Get the tags applied by an individual user to an artist on Last.fm.
Implementation of last.fm's \emph{artist.getTags} API method
(\url{http://www.last.fm/api/show/artist.getTags})
}
\examples{
artist_getTags("Hey Marseilles", "platyjus")
}
