% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist.R
\name{artist_getTopAlbums}
\alias{artist_getTopAlbums}
\title{Get the top albums for an artist.}
\usage{
artist_getTopAlbums(artist, mbid = NA, autocorrect = NA, page = NA,
  limit = NA)
}
\arguments{
\item{artist}{The artist name.}

\item{mbid}{The musicbrainz id for the artist.}

\item{autocorrect}{Transform misspelled artist names into correct artist
names, returning the correct version instead. The corrected artist name will
be returned in the response. [0|1]}

\item{page}{The page number to fetch. Defaults to first page.}

\item{limit}{The number of results to fetch per page. Defaults to 50.}
}
\value{
A list of albums.
}
\description{
Get the top albums for an artist on Last.fm, ordered by popularity.
Implementation of last.fm's \emph{artist.getTopAlbums} API method
(\url{http://www.last.fm/api/show/artist.getTopAlbums})
}
\examples{
artist_getTopAlbums("Andrew Bird")
}
