% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist.R
\name{artist_getTopTags}
\alias{artist_getTopTags}
\title{Get the top tags for an artist.}
\usage{
artist_getTopTags(artist, mbid = NA, autocorrect = NA)
}
\arguments{
\item{artist}{The artist name.}

\item{mbid}{The musicbrainz id for the artist.}

\item{autocorrect}{Transform misspelled artist names into correct artist
names, returning the correct version instead. The corrected artist name will
be returned in the response. [0|1]}
}
\value{
A list of tags.
}
\description{
Get the top tags for an artist on Last.fm, ordered by popularity.
Implementation of last.fm's \emph{artist.getTopTags} API method
(\url{http://www.last.fm/api/show/artist.getTopTags})
}
\examples{
\dontrun{
artist_getTopTags("SZA")
}
}
