% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist.R
\name{artist_getTopTracks}
\alias{artist_getTopTracks}
\title{Get the top tracks by an artist.}
\usage{
artist_getTopTracks(artist, mbid = NA, autocorrect = NA, page = NA,
  limit = NA)
}
\arguments{
\item{artist}{The artist name.}

\item{mbid}{The musicbrainz id for the artist}

\item{autocorrect}{Transform misspelled artist names into correct artist
names, returning the correct version instead. The corrected artist name will
be returned in the response. [0|1]}

\item{page}{The page number to fetch. Defaults to first page.}

\item{limit}{The number of results to fetch per page. Defaults to 50.}
}
\value{
A list of tracks.
}
\description{
Get the top tracks by an artist on Last.fm, ordered by popularity.
Implementation of last.fm's \emph{artist.getTopTracks} API method
(\url{http://www.last.fm/api/show/artist.getTopTracks})
}
\examples{
\dontrun{
artist_getTopTracks("The Suffers")
}
}
