% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist.R
\name{artist_search}
\alias{artist_search}
\title{Search for an artist by name.}
\usage{
artist_search(artist, limit = NA, page = NA)
}
\arguments{
\item{artist}{The artist name to search.}

\item{limit}{The number of results to fetch per page. Defaults to 30.}

\item{page}{The page number to fetch. Defaults to first page.}
}
\value{
A list of search results.
}
\description{
Search for an artist by name. Returns artist matches sorted by relevance.
Implementation of last.fm's \emph{artist.search} API method
(\url{http://www.last.fm/api/show/artist.search})
}
\examples{
artist_search("Helado Negro")
}
