% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-cache.r
\name{auth_cache}
\alias{auth_cache}
\alias{auth_as}
\alias{auth_list}
\alias{auth_save}
\alias{auth_load}
\title{Simple caching mechanism for auth secrets}
\usage{
auth_as(name)

auth_list()

auth_save(auth, name)

auth_load(name)
}
\arguments{
\item{name}{the name of the cache as a string}

\item{auth}{the authentication token as a string}
}
\value{
The previous authentication identity invisibly

invisibly returns path as string
}
\description{
Caching of authorization secrets relies on a simple rds file save/load mechanism.
}
\details{
Users primarily interact with the \code{\link[=auth_as]{auth_as()}} function to set the authentication
secrets for the session. Authentication identities are managed by the name \code{auth}.
}
\keyword{internal}
