% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-signature.R
\name{auth_signature}
\alias{auth_signature}
\alias{add_signature}
\alias{create_signature}
\alias{hash_signature}
\title{Request Signing}
\usage{
add_signature(params, auth_client)

create_signature(params, auth_client)

hash_signature(raw_signature)
}
\arguments{
\item{params}{Named list of query parameters}
}
\value{
\itemize{
\item \code{\link[=add_signature]{add_signature()}} returns a list
\item \code{\link[=create_signature]{create_signature()}} and \code{\link[=hash_signature]{hash_signature()}} return strings
}
}
\description{
Requests against the last.fm API must be signed using the following algorithm
as described \href{https://www.last.fm/api/desktopauth#_6-sign-your-calls}{in the API documenation}.

A short description of the methodology is included here for completeness:
\enumerate{
\item Sort query parameters by key
\item Create a single string with all parameters formatted as follows
\verb{<key><value>} using UTF-8 encoding,
\item Append the Client Secret value (without a key),
\item Hash the resulting string using a 32bit md5 hash
}

The signature is then added as extra parameter named \code{api_sig}.

These functions are not designed to be invoked directly. Requests are signed
by default when using \code{\link[=lastfm_api]{lastfm_api()}} and all other dependent raw and tidy
endpoint functions.
\itemize{
\item \code{\link[=add_signature]{add_signature()}} takes a list with request parameters and an auth_client
signs it by adding the \code{api_sig} element.
\item \code{\link[=create_signature]{create_signature()}} performs steps 1, 2, and 3
\item \code{\link[=hash_signature]{hash_signature()}} performs step 4
}
}
\keyword{internal}
