% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chart_getTopArtists}
\alias{chart_getTopArtists}
\title{Get the top artists chart.}
\usage{
chart_getTopArtists(limit = NA, page = NA)
}
\arguments{
\item{limit}{The number of results to fetch per page. Defaults to 50.}

\item{page}{The page number to fetch. Defaults to first page.}
}
\value{
A list of the top artists chart.
}
\description{
Get the top artists chart.
Implementation of last.fm's \emph{chart.getTopArtists} API method
(\url{http://www.last.fm/api/show/chart.getTopArtists})
}
\examples{
\dontrun{
chart_getTopArtists()
}
}
