% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chart_getTopTags}
\alias{chart_getTopTags}
\title{Get the top tags chart.}
\usage{
chart_getTopTags(limit = NA, page = NA, api_key = lastkey)
}
\arguments{
\item{limit}{The number of results to fetch per page. Defaults to 50.}

\item{page}{The page number to fetch. Defaults to first page.}

\item{api_key}{A Last.fm API key.}
}
\value{
A list of the top tags chart.
}
\description{
Get the top tags chart.
Implementation of last.fm's \emph{chart.getTopTags} API method
(\url{http://www.last.fm/api/show/chart.getTopTags})
}
\examples{
chart_getTopTags()
}

