% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastfmr-package.R
\name{common-params}
\alias{common-params}
\title{Common parameters in API methods}
\arguments{
\item{lang}{(Optional) The biography language.}

\item{autocorrect}{(Optional) Whether to correct the artist/album/track name names provided in the response. Defaults to TRUE if not provided.}
}
\description{
Common parameters in API methods
}
\section{Language Setting}{


Biographies for albums and artists are returned in English by
default. To get the biography in other languages, provide a \href{https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes}{ISO 639-1 country code}
via the \code{lang} argument of supporting functions.
}

\section{Name Autocorrection}{


Endpoints that take artist/track/album names try to auto-correct misspelled names
in the search query. This behaviour can be overriden by calling the function
with argument \code{autocorrect} set to \code{FALSE}.
}

