% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-signature.R
\name{create_signature_fragment}
\alias{create_signature_fragment}
\title{Builds a signature fragment from a list of query parameters.}
\usage{
create_signature_fragment(params)
}
\arguments{
\item{params}{Named list of query parameters}
}
\value{
String containing ordered key-value pairs of query parameters
}
\description{
This should not be involved directly. Use \code{sign_request} to sign a request
all queries are automatically signed where appropriate.
}
\seealso{
Other auth_signature: 
\code{\link{sign_request}()}
}
\concept{auth_signature}
