% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastfm.R
\name{fm_api}
\alias{fm_api}
\title{Query a Last.FM endpoint.}
\usage{
fm_api(query, .limit = NULL)
}
\arguments{
\item{query}{A list of parameters that define the query. Use `fm_query`
to create an appopriate object.}
}
\value{
A LastFM API object.
}
\description{
Query a Last.FM endpoint.
}
