% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{fm_user_get_top_tracks}
\alias{fm_user_get_top_tracks}
\title{Get the top tracks listened to by a user. You can stipulate a time period.
Sends the overall chart by default.}
\usage{
fm_user_get_top_tracks(username, period = "7day", page = 1)
}
\arguments{
\item{username}{The user name to fetch top tracks for.}

\item{period}{overall | 7day | 1month | 3month | 6month | 12month - The time
period over which to retrieve top tracks for.}

\item{page}{The page number to fetch. Defaults to first page.}
}
\value{

}
\description{
Get the top tracks listened to by a user. You can stipulate a time period.
Sends the overall chart by default.
}
