% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{geo_getTopArtists}
\alias{geo_getTopArtists}
\title{Get the most popular artists on Last.fm by country.}
\usage{
geo_getTopArtists(country, limit = NA, page = NA)
}
\arguments{
\item{country}{A country name, as defined by the ISO 3166-1 country names standard.}

\item{limit}{The number of results to fetch per page. Defaults to 50.}

\item{page}{TThe page number to fetch. Defaults to first page.}
}
\value{
A list of the artists.
}
\description{
Get the most popular artists on Last.fm by country.
Implementation of last.fm's \emph{geo.getTopArtists}
API method (\url{http://www.last.fm/api/show/geo.getTopArtists})
}
\examples{
geo_getTopArtists("Iceland")
}
