% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{geo_getTopTracks}
\alias{geo_getTopTracks}
\title{Get the most popular tracks on Last.fm last week by country.}
\usage{
geo_getTopTracks(country, location = NA, limit = NA, page = NA)
}
\arguments{
\item{country}{A country name, as defined by the ISO 3166-1 country names standard.}

\item{location}{A metro name, to fetch the charts for
(must be within the country specified)}

\item{limit}{The number of results to fetch per page. Defaults to 50.}

\item{page}{TThe page number to fetch. Defaults to first page.}
}
\value{
A list of the tracks.
}
\description{
Get the most popular tracks on Last.fm last week by country.
Implementation of last.fm's \emph{geo.getTopTracks}
API method (\url{http://www.last.fm/api/show/geo.getTopTracks})
}
\examples{
\dontrun{
geo_getTopTracks("Nigeria")
}
}
