% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artist_info.R
\name{get_artist_info}
\alias{get_artist_info}
\title{Get info about artists}
\usage{
get_artist_info(artist_vector)
}
\arguments{
\item{artist_vector}{\code{character} vector with specified artists}
}
\value{
\code{data.table} object with columns: artist, artist_tag, global_listners, global_scrobbles
}
\description{
Get info about artists
}
\examples{
artists <- c("Anthrax", "Metallica", "Megadeth", "Slayer")
artist_info <- get_artist_info(artist_vector = artists)

}
\seealso{
\code{\link{get_library_info}}
}
