% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_library_info.R
\name{get_library_info}
\alias{get_library_info}
\title{Get info about all artists from a user's library}
\usage{
get_library_info(user)
}
\arguments{
\item{user}{Last.fm username}
}
\value{
\code{data.table} object with columns: artist, artist_tag, global_listners, global_scrobbles
}
\description{
Get info about all artists from a user's library
}
\examples{
artist_info <- get_library_info(user = "enter_your_username")

}
\seealso{
\code{\link{get_library_info}}
}
