% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_similar.R
\name{get_similar}
\alias{get_similar}
\title{Get artist similarity relationships}
\usage{
get_similar(artist)
}
\arguments{
\item{artist}{\code{character} one-element vector with a specified artist}
}
\value{
\code{data.table} object structured as edgelist
}
\description{
Get artist similarity relationships
}
\examples{
# Get similarity two levels from given artist
edgelist <- get_similar('Closterkeller')
level2 <- rbindlist(lapply(
  edgelist[, To], get_similar
))
edgelist <- rbindlist(list(edgelist, level2))

}
