% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tags.R
\name{get_tags}
\alias{get_tags}
\title{Get all tags and their frequency for given artists}
\usage{
get_tags(artist_vector)
}
\arguments{
\item{artist_vector}{\code{character} vector with specified artists}
}
\value{
\code{data.table} object with columns: artist, tag, tag_freq
}
\description{
Get all tags and their frequency for given artists
}
\examples{
tags <- get_tags(artist_vector = c('Saxon', 'Iron Maiden'))

}
