% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracks.R
\name{get_tracks}
\alias{get_tracks}
\title{Get all tracks and their popularity for given artists}
\usage{
get_tracks(artist)
}
\arguments{
\item{artist}{one-element \code{character} vector with a specified artist}
}
\value{
\code{data.table} object with columns: track, listeners, scrobbles
}
\description{
Get all tracks and their popularity for given artists
}
\examples{
tracks <- get_tracks(artist = 'Siekiera')

}
