% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastfm.R
\name{lastfm_api}
\alias{lastfm_api}
\title{Query a Last.FM endpoint.}
\usage{
lastfm_api(query)
}
\arguments{
\item{query}{A list of parameters that define the query. Use `lastfm_query`
to create an appopriate object.}
}
\value{
A LastFM response object including the response from the endpoint,
  pagination information and the request.
}
\description{
Query a Last.FM endpoint.
}
