% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{lastfm_recent_tracks}
\alias{lastfm_recent_tracks}
\title{Get recently played tracks by username}
\usage{
lastfm_recent_tracks(username, page = 1)
}
\arguments{
\item{username}{}
}
\value{

}
\description{
Get recently played tracks by username
}
