% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastfmr-package.R
\docType{package}
\name{lastfmr-package}
\alias{lastfmr}
\alias{lastfmr-package}
\title{lastfmr: Work with Last.FM using R}
\description{
Query the last.fm API from R This package supports a handful of the endpoints and simplifies working with the API in a way that is more idiomatic. You can either query the raw endpoints or use the "tidy" endpoints to get nicer dataframes.
}
\keyword{internal}
