% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastfm.R
\name{lastfmr}
\alias{lastfmr}
\title{Query a Last.FM endpoint.}
\usage{
lastfmr(method, ..., .limit = 50, .page = 1)
}
\arguments{
\item{method}{string stating the endpoint to call}

\item{...}{named parameters to be passed to the endpoint}

\item{.limit}{Number of response to return in a single request, 50 by
default, the maximum allowed by the last.fm API.}

\item{.page}{Page to query, by default query the first page}
}
\value{
A LastFM API object.
}
\description{
Query a Last.FM endpoint.
}
