% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{library_getArtists}
\alias{library_getArtists}
\title{Get a list of the artists in a user's library.}
\usage{
library_getArtists(user, limit = NA, page = NA, api_key = lastkey)
}
\arguments{
\item{user}{The user whose library you want to fetch.}

\item{limit}{The number of results to fetch per page. Defaults to 50.}

\item{page}{The page number you wish to scan to.}

\item{api_key}{A Last.fm API key.}
}
\value{
A list of the artists.
}
\description{
A paginated list of all the artists in a user's library,
with play counts and tag counts.
Implementation of last.fm's \emph{library.getArtists} API method
(\url{http://www.last.fm/api/show/library.getArtists})
}
\examples{
library_getArtists("platyjus")
}
