% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{makeCall}
\alias{makeCall}
\title{Perform an HTTP GET Request.}
\usage{
makeCall(query = NULL)
}
\arguments{
\item{query}{List of query parameters.}
}
\value{
An HTTP \code{\link[httr]{response}}
}
\description{
Perform an HTTP GET Request.
}
