% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-user.R
\name{match_period}
\alias{match_period}
\title{Validate period}
\usage{
match_period(period = "7day")
}
\arguments{
\item{period}{(Optional) The time period to aggregate tracks/albums/artists over as a string.
One of overall, 7day, 1month, 3month, 6month, 12month. \verb{7day} by default.}
}
\description{
Validate period
}
\keyword{internal}
