% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{match_tagging_type}
\alias{match_tagging_type}
\title{Validate tagging type}
\usage{
match_tagging_type(tagging_type)
}
\arguments{
\item{tagging_type}{The objects on which the tags were applied as a string}
}
\description{
Validate tagging type
}
\keyword{internal}
