% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagination.R
\name{paginated-endpoint}
\alias{paginated-endpoint}
\title{Documenting a generic paginated endpoint}
\arguments{
\item{method}{the method to query}

\item{result_node}{a string with the name of record in the response}

\item{.start_page}{the page to start pulling from, by default starts from the
1st page}

\item{.n_pages}{the total number of pages to pull. If no value is provided,
all pages are retrieved.}
}
\description{
Documenting a generic paginated endpoint
}
