% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{pipe}
\alias{pipe}
\alias{\%>\%}
\alias{\%<>\%}
\title{Pipe operator}
\description{
See \code{magrittr::\link[magrittr]{\%>\%}} or
\code{magrittr::\link[magrittr]{\%<>\%}} for details.
}
\keyword{internal}
