% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pullUserData}
\alias{pullUserData}
\title{Pull all Last.fm data for a user}
\usage{
pullUserData(api_key, user)
}
\arguments{
\item{api_key}{(apiAccount; Required) : A Last.fm API key.}

\item{user}{(apiUser; Optional) : The username for the context of the
request. If supplied, the user's playcount for this album is included
in the response.}
}
\value{
A \code{\link[tibble]{tibble}} of results.
}
\description{
Pull all Last.fm data for a user
}
