% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album.R
\name{raw_album_get_info}
\alias{raw_album_get_info}
\title{Get album information}
\usage{
raw_album_get_info(
  artist = NULL,
  album = NULL,
  mbid = NULL,
  autocorrect = TRUE,
  username = NULL,
  lang = NULL
)
}
\arguments{
\item{artist}{Required unless mbid is provided. The artist name as a string.}

\item{album}{Required unless mbid is provided. The album name as a string.}

\item{mbid}{(Optional) MusicBrainz ID for the album as a string. Can be provided instead of artist and album names}

\item{autocorrect}{(Optional) Whether to correct the artist name. Defaults to TRUE if not provided.}

\item{username}{(Optional) Username as a string}

\item{lang}{(Optional) The language to return the biography in.}
}
\value{
A vector of LastFM API objects
}
\description{
The album can be retrieved by:
1. Artist and album name OR
2. MusicBrainz ID
}
\details{
One and only one of the above must be provided. If both are provided this
defaults to using the artist and album name.

If the artist name provided is misspelled in the search query, this endpoint
returns the correct name by default. This can be overriden by calling the function
with function argument \code{autocorrect} set to \code{FALSE}.

Album play count is added to the result if the \code{username} argument is provided.

The biography language is English by default but can be changed using the \code{lang}
argument which accepts an \href{https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes}{ISO 639-1 country code}.
}
