% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album.R
\name{raw_album_search}
\alias{raw_album_search}
\title{Search for album by name}
\usage{
raw_album_search(album, .start_page = 1, .n_pages = NULL)
}
\arguments{
\item{album}{Album name as a string}

\item{.start_page}{the page to start pulling from, by default starts from the
1st page}

\item{.n_pages}{the total number of pages to pull. If no value is provided,
all pages are retrieved.}
}
\value{
A vector of LastFM API objects
}
\description{
Search for album by name
}
