% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artist.R
\name{raw_artist_search}
\alias{raw_artist_search}
\title{Search for an artist by name}
\usage{
raw_artist_search(artist, .start_page = 1, .n_pages = NULL)
}
\arguments{
\item{artist}{Artist name as a string.}

\item{.start_page}{the page to start pulling from, by default starts from the
1st page}

\item{.n_pages}{the total number of pages to pull. If no value is provided,
all pages are retrieved.}
}
\description{
Returns artist matches sorted by relevance.
}
