% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{raw_chart_get_top_artists}
\alias{raw_chart_get_top_artists}
\title{Title}
\usage{
raw_chart_get_top_artists(.page = 1, .n_pages = NULL)
}
\arguments{
\item{.n_pages}{the total number of pages to pull. If no value is provided,
all pages are retrieved.}
}
\value{
A list of lastfm objects
}
\description{
Title
}
