% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-track.R
\name{raw_track_get_info}
\alias{raw_track_get_info}
\title{Get track information}
\usage{
raw_track_get_info(
  track = NULL,
  artist = NULL,
  mbid = NULL,
  autocorrect = TRUE,
  username = NULL,
  lang = NULL
)
}
\arguments{
\item{track}{Required unless mbid is provided. The track name as a string.}

\item{artist}{Required unless mbid is provided. The artist name as a string.}

\item{mbid}{(Optional) MusicBrainz ID for the track as a string. Can be provided instead of artist and track names}

\item{autocorrect}{(Optional) Whether to correct the artist/album/track name names provided in the response. Defaults to TRUE if not provided.}

\item{username}{(Optional) Username as a string}

\item{lang}{(Optional) The biography language.}
}
\value{
A lastfmr API object, see \code{\link[=lastfmr]{lastfmr()}} for information.
}
\description{
The track can be retrieved by:
1. Artist and track name OR
2. MusicBrainz ID
}
\details{
One and only one of the above mu?t be provided. If both are provided this
defaults to using the artist and track name.

If the artist and track names provided are misspelled in the search query, this endpoint
returns the correct name by default. This can be overriden by calling the function
with function argument \code{autocorrect} set to \code{FALSE}.

Album play count is added to the result if the \code{username} argument is provided.

The biography language is English by default but can be changed using the \code{lang}
argument which accepts an \href{https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes}{ISO 639-1 country code}.
}
\section{Language Setting}{


Biographies for albums and artists are returned in English by
default. To get the biography in other languages, provide a \href{https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes}{ISO 639-1 country code}
via the \code{lang} argument of supporting functions.
}

