% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{raw_track_get_similar}
\alias{raw_track_get_similar}
\title{Search for similar tracks}
\usage{
raw_track_get_similar(
  track = NULL,
  artist = NULL,
  mbid = NULL,
  autocorrect = TRUE,
  lang = NULL
)
}
\arguments{
\item{track}{Required unless mbid is provided. The track name as a string.}

\item{artist}{Required unless mbid is provided. The artist name as a string.}

\item{mbid}{(Optional) MusicBrainz ID for the track as a string. Can be provided instead of artist and track names}

\item{autocorrect}{(Optional) Whether to correct the artist and track names. Defaults to TRUE if not provided.}

\item{lang}{(Optional) The language to return the biography in.}
}
\value{
A lastfmr API object, see \code{\link[=lastfmr]{lastfmr()}} for information.
}
\description{
The search is performed by providing a source track by:
}
\details{
\preformatted{1. Artist and track name OR
2. MusicBrainz ID
}

One and only one of the above must be provided. If both are provided this
defaults to using the artist and track name.

If the artist and track names provided are misspelled in the search query, this endpoint
returns the correct name by default. This can be overriden by calling the function
with function argument \code{autocorrect} set to \code{FALSE}.

The biography language is English by default but can be changed using the \code{lang}
argument which accepts an \href{https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes}{ISO 639-1 country code}.
}
