% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{raw_user_get_friends}
\alias{raw_user_get_friends}
\title{Title}
\usage{
raw_user_get_friends(
  username,
  recent_tracks = TRUE,
  .start_page = 1,
  .n_pages = NULL
)
}
\arguments{
\item{username}{}

\item{recent_tracks}{}

\item{.start_page}{The page to start pulling from, by default starts from the
1st page}

\item{.n_pages}{The total number of pages to pull. If no value is provided,
all pages are retrieved.}
}
\description{
Title
}
\seealso{
Other raw user functions: 
\code{\link{raw_user_get_info}()},
\code{\link{raw_user_get_loved_tracks}()},
\code{\link{raw_user_get_personal_tags}()},
\code{\link{raw_user_get_recent_tracks}()},
\code{\link{raw_user_get_top_albums}()},
\code{\link{raw_user_get_top_artists}()},
\code{\link{raw_user_get_top_tags}()},
\code{\link{raw_user_get_top_tracks}()}
}
\concept{raw user functions}
