% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{raw_user_get_info}
\alias{raw_user_get_info}
\title{Title}
\usage{
raw_user_get_info(username = NULL, .page = 1)
}
\arguments{
\item{username}{}

\item{.page}{Page to query, by default query the first page}
}
\value{
A lastfmr API object, see \code{\link[=lastfmr]{lastfmr()}} for information.
}
\description{
Title
}
\seealso{
Other raw user functions: 
\code{\link{raw_user_get_friends}()},
\code{\link{raw_user_get_loved_tracks}()},
\code{\link{raw_user_get_personal_tags}()},
\code{\link{raw_user_get_recent_tracks}()},
\code{\link{raw_user_get_top_albums}()},
\code{\link{raw_user_get_top_artists}()},
\code{\link{raw_user_get_top_tags}()},
\code{\link{raw_user_get_top_tracks}()}
}
\concept{raw user functions}
