% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{raw_user_get_personal_tags}
\alias{raw_user_get_personal_tags}
\title{Get items that were tagged by a user with a specific tag}
\usage{
raw_user_get_personal_tags(
  username,
  tag,
  tagging_type,
  .start_page = 1,
  .n_pages = NULL
)
}
\arguments{
\item{username}{The username as a string}

\item{tag}{The tag you're interested in.}

\item{tagging_type}{A string, see description above for accepted types.}

\item{.start_page}{the page to start pulling from, by default starts from the
1st page}

\item{.n_pages}{the total number of pages to pull. If no value is provided,
all pages are retrieved.}
}
\description{
Items are queried by type using \code{tagging_type} which can take one of
(artist, album, track).
}
\seealso{
Other raw user functions: 
\code{\link{raw_user_get_friends}()},
\code{\link{raw_user_get_info}()},
\code{\link{raw_user_get_loved_tracks}()},
\code{\link{raw_user_get_recent_tracks}()},
\code{\link{raw_user_get_top_albums}()},
\code{\link{raw_user_get_top_artists}()},
\code{\link{raw_user_get_top_tags}()},
\code{\link{raw_user_get_top_tracks}()}
}
\concept{raw user functions}
