% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{raw_user_get_recent_tracks}
\alias{raw_user_get_recent_tracks}
\title{Get a list of the recent tracks listened to by this user.}
\usage{
raw_user_get_recent_tracks(
  username,
  from = NULL,
  to = NULL,
  extended = 0,
  .start_page = 1,
  .n_pages = NULL,
  .limit = 50
)
}
\arguments{
\item{username}{The last.fm username to fetch the recent tracks of.}

\item{from}{(optional) Beginning timestamp of a range - only display
scrobbles after this time.}

\item{to}{(optional) End timestamp of a range - only display scrobbles before
this time.}

\item{extended}{(optional) Whether to include extended data in each artist, and whether
or not the user has loved each track.}

\item{.start_page}{The page to start pulling from, by default starts from the
1st page}

\item{.n_pages}{The total number of pages to pull. If no value is provided,
all pages are retrieved.}

\item{.limit}{Total number of results to include in a single page. 50 by default.}
}
\value{
A vector of LastFM API objects
}
\description{
Get a list of the recent tracks listened to by this user.
}
\seealso{
Other raw functions: 
\code{\link{raw_user_get_top_albums}()},
\code{\link{raw_user_get_top_artists}()},
\code{\link{raw_user_get_top_tracks}()}
}
\concept{raw functions}
