% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{raw_user_get_top_tags}
\alias{raw_user_get_top_tags}
\title{Get top tags used by a user}
\usage{
raw_user_get_top_tags(username, .start_page = 1, .n_pages = NULL)
}
\arguments{
\item{username}{Username as a string}

\item{.start_page}{the page to start pulling from, by default starts from the
1st page}

\item{.n_pages}{the total number of pages to pull. If no value is provided,
all pages are retrieved.}
}
\value{

}
\description{
Get top tags used by a user
}
\seealso{
Other raw user functions: 
\code{\link{raw_user_get_friends}()},
\code{\link{raw_user_get_info}()},
\code{\link{raw_user_get_loved_tracks}()},
\code{\link{raw_user_get_personal_tags}()},
\code{\link{raw_user_get_recent_tracks}()},
\code{\link{raw_user_get_top_albums}()},
\code{\link{raw_user_get_top_artists}()},
\code{\link{raw_user_get_top_tracks}()}
}
\concept{raw user functions}
