% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth-signature.R
\name{sign_request}
\alias{sign_request}
\title{Adds an encrypted signature to a query prior to request.}
\usage{
sign_request(params)
}
\arguments{
\item{params}{Named list of query parameters}
}
\value{
String containing ordered key-value pairs of query parameters
}
\description{
This should not be invoked directly by end-users and instead prefer to use
\code{\link[=lastfm_api]{lastfm_api()}} or one of the \code{raw} or \verb{tidy data} functions.
}
\seealso{
Other auth_signature: 
\code{\link{create_signature_fragment}()}
}
\concept{auth_signature}
