% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastfm.R
\name{single-page-endpoint}
\alias{single-page-endpoint}
\title{Last.FM querying endpoint}
\arguments{
\item{method}{The name of the API method \href{https://www.last.fm/api}{found here}.}

\item{result_node}{The node in the response that contains the data.}

\item{.page}{Page to query, by default query the first page}
}
\value{
A \link{lastfm} API object.
}
\description{
Used for documenting single-query endpoints
}
