% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track_getCorrection}
\alias{track_getCorrection}
\title{Get corrections of a track title or artist.}
\usage{
track_getCorrection(track, artist)
}
\arguments{
\item{track}{The track name to correct.}

\item{artist}{The artist name to correct.}
}
\value{
A list of corrections.
}
\description{
Use the last.fm corrections data to check whether the supplied track has a
correction to a canonical track
Implementation of last.fm's \emph{track.getCorrection} API method
(\url{http://www.last.fm/api/show/track.getCorrection})
}
\examples{
track_getCorrection("Mrbrownstone", "guns and roses")
}
