% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track_getInfo}
\alias{track_getInfo}
\title{Get the metadata for a track.}
\usage{
track_getInfo(track, artist, mbid = NA, autocorrect = NA, username = NA,
  api_key = lastkey)
}
\arguments{
\item{track}{The track name.}

\item{artist}{The artist name.}

\item{mbid}{The musicbrainz id for the track}

\item{autocorrect}{Transform misspelled artist and track names into correct artist and track names,
returning the correct version instead.
The corrected artist and track name will be returned in the response. [0|1]}

\item{username}{The username for the context of the request.
If supplied, the user's playcount for this track and
whether they have loved the track is included in the response.}

\item{api_key}{A Last.fm API key.}
}
\value{
A list of the metadata for a track.
}
\description{
Get the metadata for a track on Last.fm using the artist/track name or a musicbrainz id.
Implementation of last.fm's \emph{track.getInfo} API method
(\url{http://www.last.fm/api/show/track.getInfo})
}
\examples{
track_getInfo("King Kunta", "Kendrick Lamar")
}
