% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track_getTags}
\alias{track_getTags}
\title{Get the tags applied by an individual user to a track.}
\usage{
track_getTags(track, artist, mbid = NA, autocorrect = NA, username = NA,
  api_key = lastkey)
}
\arguments{
\item{track}{The track name.}

\item{artist}{The artist name.}

\item{mbid}{The musicbrainz id for the track}

\item{autocorrect}{Transform misspelled artist and track names into correct
artist and track names, returning the correct version instead.
The corrected artist and track name will be returned in the response. [0|1]}

\item{username}{If called in non-authenticated mode you must specify the user
to look up.}

\item{api_key}{A Last.fm API key.}
}
\value{
A list of the metadata for a track.
}
\description{
Get the tags applied by an individual user to a track on Last.fm.
Implementation of last.fm's \emph{track.getTags} API method
(\url{http://www.last.fm/api/show/track.getTags})
}
\examples{
track_getTags("Sleeping In", "The Postal Service")
}

