% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track_getTopTags}
\alias{track_getTopTags}
\title{Get the top tags for this track.}
\usage{
track_getTopTags(track, artist, mbid = NA, autocorrect = NA,
  api_key = lastkey)
}
\arguments{
\item{track}{The track name.}

\item{artist}{The artist name.}

\item{mbid}{The musicbrainz id for the track}

\item{autocorrect}{Transform misspelled artist and track names into correct
artist and track names, returning the correct version instead.
The corrected artist and track name will be returned in the response. [0|1]}

\item{api_key}{A Last.fm API key.}
}
\value{
A list of the top tags for a track.
}
\description{
Get the top tags for this track on Last.fm, ordered by tag count.
Supply either track & artist name or mbid.
Implementation of last.fm's \emph{track.getTopTags} API method
(\url{http://www.last.fm/api/show/track.getTopTags})
}
\examples{
track_getTopTags("Tommy Chong", "The Blue Scholars")
}

