% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track_search}
\alias{track_search}
\title{Search for a track by track name.}
\usage{
track_search(track, artist = NA, limit = NA, page = NA,
  api_key = lastkey)
}
\arguments{
\item{track}{The track name.}

\item{artist}{Narrow your search by specifying an artist.}

\item{limit}{The number of results to fetch per page. Defaults to 30.}

\item{page}{The page number to fetch. Defaults to first page.}

\item{api_key}{A Last.fm API key.}
}
\value{
A list of track matches sorted by relevance.
}
\description{
Search for a track by track name. Returns track matches sorted by relevance.
Supply either track & artist name or mbid.
Implementation of last.fm's \emph{track.search} API method
(\url{http://www.last.fm/api/show/track.search})
}
\examples{
track_search("Depreston")
}
