% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user}
\alias{user}
\title{S3 generics for user API calls}
\arguments{
\item{api_key}{(apiAccount; Required) : A Last.fm API key.}

\item{user}{(apiUser; Required) : The user name to fetch top albums for.}

\item{period}{(string; Optional) : The time period over which to retrieve
top albums for. Must be one of
\code{{overall, 7day, 1month, 3month, 6month, 12month}}}

\item{artist}{(character; Required unless mbid) : The artist name.}

\item{album}{(character; Required unless mbid) : The album name.}

\item{limit}{(int; Optional) : The number of results to fetch per page.
Defaults to 50.}

\item{page}{(int; Optional) : The page number to fetch. Defaults to first
page.}
}
\value{
A \code{\link[tibble]{tibble}} of results
}
\description{
S3 generics for user API calls
}
\seealso{
Other user: \code{\link{userGetArtistTracks}},
  \code{\link{userGetTopAlbums}},
  \code{\link{userGetTopArtists}},
  \code{\link{userGetTopTracks}}
}
