% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user_get_recent_tracks}
\alias{user_get_recent_tracks}
\title{Get a list of the recent tracks listened to by this user as tidy data}
\usage{
user_get_recent_tracks(
  username,
  from = NULL,
  to = NULL,
  extended = 0,
  .start_page = 1,
  .n_pages = NULL
)
}
\arguments{
\item{username}{The last.fm username to fetch the recent tracks of.}

\item{from}{(optional) Beginning timestamp of a range - only display
scrobbles after this time.}

\item{to}{(optional) End timestamp of a range - only display scrobbles before
this time.}

\item{extended}{(optional) Whether to include extended data in each artist, and whether
or not the user has loved each track.}

\item{.start_page}{The page to start pulling from, by default starts from the
1st page}

\item{.n_pages}{The total number of pages to pull. If no value is provided,
all pages are retrieved.}
}
\value{
A tibble
}
\description{
Get a list of the recent tracks listened to by this user as tidy data
}
\seealso{
Other tidy user functions: 
\code{\link{user_get_top_albums}()},
\code{\link{user_get_top_artists}()}
}
\concept{tidy user functions}
