% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user_get_top_albums}
\alias{user_get_top_albums}
\title{Get the top albums by scrobbles for a given user as tidy data}
\usage{
user_get_top_albums(
  username,
  period = "7day",
  .start_page = 1,
  .n_pages = NULL
)
}
\arguments{
\item{username}{The user name to fetch top albums for.}

\item{period}{(optional) overall | 7day | 1month | 3month | 6month | 12month - The time period over which to retrieve top albums for.}

\item{.start_page}{the page to start pulling from, by default starts from the
1st page}

\item{.n_pages}{the total number of pages to pull. If no value is provided,
all pages are retrieved.}
}
\value{
a tibble
}
\description{
You can stipulate a time period. Sends the overall chart by default.
}
\seealso{
Other tidy user functions: 
\code{\link{user_get_recent_tracks}()},
\code{\link{user_get_top_artists}()}
}
\concept{tidy user functions}
