% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user_getArtistTracks}
\alias{user_getArtistTracks}
\title{Get a list of tracks by a given artist scrobbled by this user.}
\usage{
user_getArtistTracks(user, artist, startTimestamp = NA, endTimestamp = NA,
  page = NA, api_key = lastkey)
}
\arguments{
\item{user}{The last.fm username to fetch the recent tracks of.}

\item{artist}{The artist name you are interested in.}

\item{startTimestamp}{An unix timestamp to start at.}

\item{endTimestamp}{An unix timestamp to end at.}

\item{page}{The page number to fetch. Defaults to first page.}

\item{api_key}{A Last.fm API key.}
}
\value{
A list of tracks by a given artist scrobbled by this user.
}
\description{
Get a list of tracks by a given artist scrobbled by this user,
including scrobble time. Can be limited to specific timeranges, defaults to all time.
Implementation of last.fm's \emph{user.getArtistTracks} API method
(\url{http://www.last.fm/api/show/user.getArtistTracks}).
}
\examples{
user_getArtistTracks("platyjus", "Leon Bridges")
}
