% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user_getFriends}
\alias{user_getFriends}
\title{Get a list of the user's friends.}
\usage{
user_getFriends(user, recenttracks = NA, limit = NA, page = NA,
  api_key = lastkey)
}
\arguments{
\item{user}{The last.fm username to fetch the friends of.}

\item{recenttracks}{Whether or not to include information about friends' recent listening in the response.}

\item{limit}{The number of results to fetch per page. Defaults to 50.}

\item{page}{The page number to fetch. Defaults to first page.}

\item{api_key}{A Last.fm API key.}
}
\value{
A list of the user's friends.
}
\description{
Get a list of the user's friends on Last.fm.
Implementation of last.fm's \emph{user.getFriends} API method
(\url{http://www.last.fm/api/show/user.getFriends}).
}
\examples{
user_getFriends("platyjus")
}
