% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user_getLovedTracks}
\alias{user_getLovedTracks}
\title{Get the last 50 tracks loved by a user.}
\usage{
user_getLovedTracks(user, limit = NA, page = NA)
}
\arguments{
\item{user}{The user name to fetch the loved tracks for.}

\item{limit}{The number of results to fetch per page. Defaults to 50.}

\item{page}{The page number to fetch. Defaults to first page.}
}
\value{
A list of the last 50 tracks loved by a user.
}
\description{
Get the last 50 tracks loved by a user.
Implementation of last.fm's \emph{user.getLovedTracks} API method
(\url{http://www.last.fm/api/show/user.getLovedTracks}).
}
\examples{
\dontrun{
user_getLovedTracks("platyjus")
}
}
