% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user_getPersonalTags}
\alias{user_getPersonalTags}
\title{Get the user's personal tags.}
\usage{
user_getPersonalTags(user, tag, taggingtype, limit = NA, page = NA,
  api_key = lastkey)
}
\arguments{
\item{user}{The user who performed the taggings.}

\item{tag}{The tag you're interested in.}

\item{taggingtype}{The type of items which have been tagged [artist|album|track].}

\item{limit}{The number of results to fetch per page. Defaults to 50.}

\item{page}{The page number to fetch. Defaults to first page.}

\item{api_key}{A Last.fm API key.}
}
\value{
A list of the user's personal tags.
}
\description{
Get the user's personal tags.
Implementation of last.fm's \emph{user.getPersonalTags} API method
(\url{http://www.last.fm/api/show/user.getPersonalTags}).
}
\examples{
user_getPersonalTags("platyjus", "indie", "artist")
}

