% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user_getRecentTracks}
\alias{user_getRecentTracks}
\title{Get a list of the recent tracks listened to by user.}
\usage{
user_getRecentTracks(user, limit = NA, page = 1, from = NA, to = NA,
  extended = 0, api_key = lastkey)
}
\arguments{
\item{user}{The last.fm username to fetch the recent tracks of.}

\item{limit}{The number of results to fetch per page. Defaults to 50. Maximum is 200.}

\item{page}{The page number you wish to scan to.}

\item{from}{Beginning timestamp of a range - only display scrobbles after this time,
in UNIX timestamp format (integer number of seconds since 00:00:00, January 1st 1970 UTC).
This must be in the UTC time zone.}

\item{to}{End timestamp of a range - only display scrobbles before this time,
in UNIX timestamp format (integer number of seconds since 00:00:00, January 1st 1970 UTC).
This must be in the UTC time zone.}

\item{extended}{Includes extended data in each artist, and whether or not the user has loved each track}

\item{api_key}{A Last.fm API key.}
}
\value{
A list of the recent tracks.
}
\description{
Get a list of the recent tracks listened to by this user.
Also includes the currently playing track with the nowplaying="true" attribute
if the user is currently listening.
Implementation of last.fm's \emph{user.getRecentTracks} API method
(\url{http://www.last.fm/api/show/user.getRecentTracks}).
}
\examples{
user_getRecentTracks("platyjus")
}

