% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user_getTopArtists}
\alias{user_getTopArtists}
\title{Get the top artists listened to by a user.}
\usage{
user_getTopArtists(user, period = NA, limit = NA, page = NA)
}
\arguments{
\item{user}{The user name to fetch top artists for.}

\item{period}{overall | 7day | 1month | 3month | 6month | 12month -
The time period over which to retrieve top artists for.}

\item{limit}{The number of results to fetch per page. Defaults to 50.}

\item{page}{The page number to fetch. Defaults to first page.}
}
\value{
A list of the top artists listened to by a user.
}
\description{
Get the top artists listened to by a user. You can stipulate a time period.
Sends the overall chart by default.
Implementation of last.fm's \emph{user.getTopArtists} API method
(\url{http://www.last.fm/api/show/user.getTopArtists}).
}
\examples{
\dontrun{
user_getTopArtists("platyjus")
}
}
