% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user_getTopTags}
\alias{user_getTopTags}
\title{Get the top tags listened to by a user.}
\usage{
user_getTopTags(user, limit = NA, api_key = lastkey)
}
\arguments{
\item{user}{The user name.}

\item{limit}{Limit the number of tags returned}

\item{api_key}{A Last.fm API key.}
}
\value{
A list of the top tags listened to by a user.
}
\description{
Get the top tags listened to by a user.
Implementation of last.fm's \emph{user.getTopTags} API method
(\url{http://www.last.fm/api/show/user.getTopTags}).
}
\examples{
user_getTopTags("platyjus")
}
