% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user_getTopTracks}
\alias{user_getTopTracks}
\title{Get the top tracks listened to by a user.}
\usage{
user_getTopTracks(user, period = NA, limit = NA, page = NA,
  api_key = lastkey)
}
\arguments{
\item{user}{The user name to fetch top tracks for.}

\item{period}{overall | 7day | 1month | 3month | 6month | 12month -
The time period over which to retrieve top tracks for.}

\item{limit}{The number of results to fetch per page. Defaults to 50.}

\item{page}{The page number to fetch. Defaults to first page.}

\item{api_key}{A Last.fm API key.}
}
\value{
A list of the top tracks listened to by a user.
}
\description{
Get the top tracks listened to by a user. You can stipulate a time period.
Sends the overall chart by default.
Implementation of last.fm's \emph{user.getTopTracks} API method
(\url{http://www.last.fm/api/show/user.getTopTracks}).
}
\examples{
user_getTopTracks("platyjus")
}
