% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user_getWeeklyArtistChart}
\alias{user_getWeeklyArtistChart}
\title{Get an artist chart for a user profile.}
\usage{
user_getWeeklyArtistChart(user, from = NA, to = NA)
}
\arguments{
\item{user}{The last.fm username to fetch the charts of.}

\item{from}{The date at which the chart should start from.}

\item{to}{The date at which the chart should end on.}
}
\value{
A list of an artist chart for a user profile.
}
\description{
Get an artist chart for a user profile, for a given date range.
If no date range is supplied, it will return the most recent artist chart for this user.
Implementation of last.fm's \emph{user.getWeeklyArtistChart} API method
(\url{http://www.last.fm/api/show/user.getWeeklyArtistChart}).
}
\examples{
user_getWeeklyArtistChart("platyjus")
}
