% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user_getWeeklyChartList}
\alias{user_getWeeklyChartList}
\title{Get a list of available charts for this user.}
\usage{
user_getWeeklyChartList(user)
}
\arguments{
\item{user}{TThe last.fm username to fetch the charts list for.}
}
\value{
A list of an artist chart for a user profile.
}
\description{
Get a list of available charts for this user,
expressed as date ranges which can be sent to the chart services.
Implementation of last.fm's \emph{user.getWeeklyChartList} API method
(\url{http://www.last.fm/api/show/user.getWeeklyChartList}).
}
\examples{
user_getWeeklyChartList("platyjus")
}
