% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{userGetArtistTracks}
\alias{userGetArtistTracks}
\title{Get a list of tracks scrobbled by a given last.fm user}
\usage{
userGetArtistTracks(artist, api_key, user, limit = NULL, page = NULL)
}
\arguments{
\item{artist}{(character; Required unless mbid) : The artist name.}

\item{api_key}{(apiAccount; Required) : A Last.fm API key.}

\item{user}{(apiUser; Required) : The user name to fetch top albums for.}

\item{limit}{(int; Optional) : The number of results to fetch per page.
Defaults to 50.}

\item{page}{(int; Optional) : The page number to fetch. Defaults to first
page.}
}
\value{
A \code{\link[tibble]{tibble}} of results
}
\description{
Get a list of tracks by a given artist scrobbled by this user, including
scrobble time. Can be limited to specific timeranges, defaults to all
time.
}
\references{
See \url{https://www.last.fm/api/show/user.getArtistTracks}
}
\seealso{
Other user: \code{\link{userGetTopAlbums}},
  \code{\link{userGetTopArtists}},
  \code{\link{userGetTopTracks}}, \code{\link{user}}
}
