% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{userGetTopAlbums}
\alias{userGetTopAlbums}
\title{Get Albums with Most Plays for User}
\usage{
userGetTopAlbums(api_key, user, period = c("overall", "7day", "1month",
  "3month", "6month", "12month"), limit = NULL, page = NULL)
}
\arguments{
\item{api_key}{(string; Required) : A Last.fm API key.}

\item{user}{(string; Required) : The user name to fetch top albums for.}

\item{period}{(string; Optional) : The time period over which to retrieve
top albums for. Must be one of
\code{{overall, 7day, 1month, 3month, 6month, 12month}}}

\item{limit}{(int; Optional) : The number of results to fetch per page.
Defaults to 50.}

\item{page}{(int; Optional) : The page number to fetch. Defaults to first
page.}
}
\value{
parsed JSON
}
\description{
Get Albums with Most Plays for User
}
