% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{userGetTopArtists}
\alias{userGetTopArtists}
\title{Get the top artists listened to by a user}
\usage{
userGetTopArtists(api_key, user, period = c("overall", "7day", "1month",
  "3month", "6month", "12month"), limit = NULL, page = NULL)
}
\arguments{
\item{api_key}{(apiAccount; Required) : A Last.fm API key.}

\item{user}{(apiUser; Required) : The user name to fetch top albums for.}

\item{period}{(string; Optional) : The time period over which to retrieve
top albums for. Must be one of
\code{{overall, 7day, 1month, 3month, 6month, 12month}}}

\item{limit}{(int; Optional) : The number of results to fetch per page.
Defaults to 50.}

\item{page}{(int; Optional) : The page number to fetch. Defaults to first
page.}
}
\value{
A \code{\link[tibble]{tibble}} of results
}
\description{
Get the top artists listened to by a user. You can stipulate a time period.
Sends the overall chart by default.
}
\references{
See \url{https://www.last.fm/api/show/user.getArtistTracks}
}
