% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album.R
\name{album_getTags}
\alias{album_getTags}
\title{Get the tags applied by an user to an album.}
\usage{
album_getTags(artist, album, mbid = NA, autocorrect = NA, username = NA,
  api_key = lastkey)
}
\arguments{
\item{artist}{The artist name.}

\item{album}{The album name.}

\item{mbid}{The musicbrainz id for the album.}

\item{autocorrect}{Transform misspelled artist names into correct artist names,
returning the correct version instead.
The corrected artist name will be returned in the response. [0|1]}

\item{api_key}{A Last.fm API key.}

\item{user}{The username for the context of the request.
If supplied, the user's playcount for this album is included in the response.}
}
\value{
A list of the tags applied by an user to an album.
}
\description{
Get the tags applied by an individual user to an album on Last.fm.
To retrieve the list of top tags applied to an album by all users use album_getTopTags.
Implementation of last.fm's \emph{album.getTags} API method
(\url{http://www.last.fm/api/show/album.getTags})
}
\examples{
album_getTags("Sufjan Stevens", "Carrie & Lowell")
}

