% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth}
\alias{lastfm_auth}
\alias{fetch_request_token}
\alias{request_user_auth}
\alias{fetch_web_session}
\title{Access Token Management}
\usage{
lastfm_auth(api_key = fm_api_key())

fetch_request_token(api_key = fm_api_key())

request_user_auth(request_token, api_key = fm_api_key())

fetch_web_session(token, api_key = fm_api_key())
}
\arguments{
\item{api_key}{(optional) Application API Key}

\item{request_token}{String containing the request token obtained with \code{\link[=fetch_request_token]{fetch_request_token()}}}
}
\value{
\itemize{
\item \code{fetch_request_token()} and \code{fetch_web_session()} return strings
}

Request token


}
\description{
lastfmr implements the desktop application authentication method specified
in the \href{https://www.last.fm/api/desktopauth}{API documentation}.

By default, you are directed to a web browser, asked to sign in to your last.fm
account, and to grant lastfmr permissions to operate on your behalf.

The 1st part of the user authentication process. This token expires after
60 minutes.
}
\section{Authentication Workflow}{
The authentication workflow is execute by a call to \code{\link[=lastfm_auth]{lastfm_auth()}} which invokes
the full workflow:
\enumerate{
\item \code{\link[=fetch_request_token]{fetch_request_token()}} creates a request token for the specific API request, this authenticates a
particular API request and lasts for 60 minutes.
\item \code{\link[=request_user_auth]{request_user_auth()}} launches a browser to request end-user authorization
for access to user account data through the browser. The token created in
the first step is provided as part of the user authorization request to
associate the Application API Key with the user account providing
authorization.
\item \code{\link[=fetch_web_session]{fetch_web_session()}} fetches a session token to be used with all
end-user requests. This is done by providing the application API key and
the request token. The endpoint returns an encrypted session token that
can be attached to all user requests.
}
}

\examples{

}
\concept{auth}
