% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-chart.R
\name{chart-methods}
\alias{chart-methods}
\alias{raw_chart_get_top_artists}
\alias{chart_get_top_artists}
\alias{raw_chart_get_top_tags}
\alias{chart_get_top_tags}
\alias{raw_chart_get_top_tracks}
\alias{chart_get_top_tracks}
\title{Chart Methods}
\usage{
raw_chart_get_top_artists(.start_page = 1, .n_pages = NULL)

chart_get_top_artists(.start_page = 1, .n_pages = NULL)

raw_chart_get_top_tags(.start_page = 1, .n_pages = NULL)

chart_get_top_tags(.start_page = 1, .n_pages = NULL)

raw_chart_get_top_tracks(.start_page = 1, .n_pages = NULL)

chart_get_top_tracks(.start_page = 1, .n_pages = NULL)
}
\arguments{
\item{.start_page}{The page to start pulling from, by default starts from the
1st page}

\item{.n_pages}{The total number of pages to pull. If no value is provided,
all pages are retrieved.}
}
\value{
\itemize{
\item All raw endpoints return a list of lastfmr objects.
\item All tidy endpoints return a Tibble.
}
}
\description{
Get last.fm overall chart rankings for tags, tracks, and artists.

\code{\link[=raw_chart_get_top_artists]{raw_chart_get_top_artists()}}, \code{\link[=raw_chart_get_top_tags]{raw_chart_get_top_tags()}},
and \code{\link[=raw_chart_get_top_tracks]{raw_chart_get_top_tracks()}} retrieve the overall top artists, tags, and tracks, respectively.
}
