% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chart_getTopTracks}
\alias{chart_getTopTracks}
\title{Get the top tracks chart.}
\usage{
chart_getTopTracks(limit = NA, page = NA, api_key = lastkey)
}
\arguments{
\item{limit}{The number of results to fetch per page. Defaults to 50.}

\item{page}{The page number to fetch. Defaults to first page.}

\item{api_key}{A Last.fm API key.}
}
\value{
A list of the top tracks chart.
}
\description{
Get the top tracks chart.
Implementation of last.fm's \emph{chart.getTopTracks} API method
(\url{http://www.last.fm/api/show/chart.getTopTracks})
}
\examples{
chart_getTopTracks()
}
