% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scrobbles.R
\name{get_scrobbles}
\alias{get_scrobbles}
\title{Get all scroblles for a given user}
\usage{
get_scrobbles(user)
}
\arguments{
\item{user}{Last.fm username}
}
\value{
\code{data.table} object with columns: date, track, artist, album
}
\description{
Get all scroblles for a given user
}
\examples{
scrobbles <- get_scrobbles(user = "enter_your_username")

}
