% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastfm.R
\name{lastfmr}
\alias{lastfmr}
\title{Query a Last.FM endpoint.}
\usage{
lastfmr(method, ..., .limit = 50, .page = 1)
}
\arguments{
\item{method}{string stating the endpoint to call}

\item{...}{named parameters to be passed to the endpoint}

\item{.limit}{limit the number of items returned by the API, maximum 50}

\item{.page}{optional parameter which page to start scanning from}
}
\value{
A LastFM API object.
}
\description{
Query a Last.FM endpoint.
}
