% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastfm.R
\name{lastfmr}
\alias{lastfmr}
\title{Query a Last.FM endpoint.}
\usage{
lastfmr(method, ..., .page = 1, result_node = NULL)
}
\arguments{
\item{method}{a string}

\item{...}{named parameters to be passed to the endpoint}

\item{.page}{Page to query, by default query the first page}

\item{result_node}{a string}
}
\value{
A LastFM API object.
}
\description{
Query a Last.FM endpoint.
}
