% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagination.R
\name{paginate}
\alias{paginate}
\title{A generic paginated request}
\usage{
paginate(method, ..., .start_page = 1, .n_pages = NULL, .limit = 50)
}
\arguments{
\item{method}{The method to query}

\item{...}{Endpoint parameters to pass to the API endpoint}

\item{.start_page}{The page to start pulling from, by default starts from the
1st page}

\item{.n_pages}{The total number of pages to pull. If no value is provided,
all pages are retrieved.}

\item{.limit}{Total number of results to include in a single page. 50 by default.}
}
\value{
A list of lastfmr objects
}
\description{
The pagination approach taken here involves constructing successive requests
from an initial response. This initial response represents the first page in
our multipage request.
}
