% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagination.R
\name{paginate}
\alias{paginate}
\title{Get a concatenated result from multipage endpoints}
\usage{
paginate(resp, n_pages = NULL)
}
\arguments{
\item{resp}{response object of type \code{lastfmr}}

\item{n_pages}{}
}
\value{

}
\description{
The pagination approach taken here involves constructing successive requests
from an initial response. This initial response represents the first page
in our multipage request.
}
