% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagination.R
\name{paginate}
\alias{paginate}
\title{A generic paginated request}
\usage{
paginate(method, result_node, ..., .start_page = 1, .n_pages = NULL)
}
\arguments{
\item{method}{the method to query}

\item{result_node}{a string with the name of record in the response}

\item{...}{Endpoint parameters to pass to the API endpoint}

\item{.start_page}{the page to start pulling from, by default starts from the
1st page}

\item{.n_pages}{the total number of pages to pull. If no value is provided,
all pages are retrieved.}
}
\value{
A list of lastfmr objects
}
\description{
The pagination approach taken here involves constructing successive requests
from an initial response. This initial response represents the first page in
our multipage request.
}
