% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagination.R
\name{paginated-endpoint}
\alias{paginated-endpoint}
\title{Documenting a generic paginated endpoint}
\arguments{
\item{method}{The method to query}

\item{.start_page}{The page to start pulling from, by default starts from the
1st page}

\item{.n_pages}{The total number of pages to pull. If no value is provided,
all pages are retrieved.}

\item{.limit}{Total number of results to include in a single page. 50 by default.}
}
\description{
Documenting a generic paginated endpoint
}
