% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-chart.R
\name{raw_chart_get_top_artists}
\alias{raw_chart_get_top_artists}
\title{Get list of top artists}
\usage{
raw_chart_get_top_artists(.start_page = 1, .n_pages = NULL)
}
\arguments{
\item{.start_page}{The page to start pulling from, by default starts from the
1st page}

\item{.n_pages}{The total number of pages to pull. If no value is provided,
all pages are retrieved.}
}
\value{
A vector of LastFM API objects
}
\description{
Get list of top artists
}
