% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-track.R
\name{raw_track_get_similar}
\alias{raw_track_get_similar}
\title{Search for similar tracks}
\usage{
raw_track_get_similar(
  track = NULL,
  artist = NULL,
  mbid = NULL,
  autocorrect = TRUE,
  lang = NULL
)
}
\arguments{
\item{track}{Required unless mbid is provided. The track name as a string.}

\item{artist}{Required unless mbid is provided. The artist name as a string.}

\item{mbid}{(Optional) MusicBrainz ID for the track as a string. Can be provided instead of artist and track names}

\item{autocorrect}{(Optional) Whether to correct the artist/album/track name names provided in the response. Defaults to TRUE if not provided.}

\item{lang}{(Optional) The biography language.}
}
\value{
A \link{lastfm} API object.
}
\description{
The search is performed by providing a source track by:
}
\details{
\preformatted{1. Artist and track name OR
2. MusicBrainz ID
}

One and only one of the above must be provided. If both are provided this
defaults to using the artist and track name.
}
\section{Language Setting}{


Biographies for albums and artists are returned in English by
default. To get the biography in other languages, provide a \href{https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes}{ISO 639-1 country code}
via the \code{lang} argument of supporting functions.
}

\section{Name Autocorrection}{


Endpoints that take artist/track/album names try to auto-correct misspelled names
in the search query. This behaviour can be overriden by calling the function
with argument \code{autocorrect} set to \code{FALSE}.
}

