% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{raw_track_search}
\alias{raw_track_search}
\title{Search for a track by track name. Returns track matches sorted by relevance.}
\usage{
raw_track_search(track, artist = NULL, page = 1)
}
\arguments{
\item{track}{The track name.}

\item{artist}{Narrow your search by specifying an artist.}

\item{page}{The page number ot fetch. Defaults to first page.}
}
\value{
A LastFM API object
}
\description{
Search for a track by track name. Returns track matches sorted by relevance.
}
