% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{raw_track_search}
\alias{raw_track_search}
\title{Search for a track by track name}
\usage{
raw_track_search(track, artist = NULL, .start_page = 1, .n_pages = NULL)
}
\arguments{
\item{track}{The track name.}

\item{artist}{Narrow your search by specifying an artist, if not provided the
endpoint searches across artists}

\item{.start_page}{The page to start pulling from, by default starts from the
1st page}

\item{.n_pages}{The total number of pages to pull. If no value is provided,
all pages are retrieved.}
}
\description{
Returns track matches sorted by relevance.
}
