% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{raw_user_get_recent_tracks}
\alias{raw_user_get_recent_tracks}
\title{Get a list of the recent tracks listened to by this user.}
\usage{
raw_user_get_recent_tracks(
  username,
  from = NULL,
  to = NULL,
  extended = 0,
  page = 1
)
}
\arguments{
\item{username}{The last.fm username to fetch the recent tracks of.}

\item{from}{(optional) Beginning timestamp of a range - only display
scrobbles after this time.}

\item{to}{(optional) End timestamp of a range - only display scrobbles before
this time.}

\item{extended}{(optional) Whether to include extended data in each artist, and whether
or not the user has loved each track.}

\item{page}{(optional) The page number to fetch. Defaults to first page.}
}
\value{
A LastFM API object
}
\description{
Get a list of the recent tracks listened to by this user.
}
\seealso{
Other raw: 
\code{\link{raw_user_get_top_albums}()},
\code{\link{raw_user_get_top_artists}()}
}
\concept{raw}
