% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{raw_user_get_top_albums}
\alias{raw_user_get_top_albums}
\title{Get the top albums listened to by a user. You can stipulate a time period. Sends the overall chart by default.}
\usage{
raw_user_get_top_albums(username, period = "7day", page = 1)
}
\arguments{
\item{username}{The user name to fetch top albums for.}

\item{period}{(optional) overall | 7day | 1month | 3month | 6month | 12month - The time period over which to retrieve top albums for.}

\item{page}{The page number to fetch. Defaults to first page.}
}
\value{
A LastFM API object
}
\description{
Get the top albums listened to by a user. You can stipulate a time period. Sends the overall chart by default.
}
\seealso{
Other raw: 
\code{\link{raw_user_get_recent_tracks}()},
\code{\link{raw_user_get_top_artists}()}
}
\concept{raw}
