% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{raw_user_get_top_albums}
\alias{raw_user_get_top_albums}
\title{Get the top albums listened to by a user. You can stipulate a time period. Sends the overall chart by default.}
\usage{
raw_user_get_top_albums(
  username,
  period = "7day",
  .start_page = 1,
  .n_pages = NULL
)
}
\arguments{
\item{username}{The user name to fetch top albums for.}

\item{period}{(optional) overall | 7day | 1month | 3month | 6month | 12month - The time period over which to retrieve top albums for.}

\item{.start_page}{The page to start pulling from, by default starts from the
1st page}

\item{.n_pages}{The total number of pages to pull. If no value is provided,
all pages are retrieved.}
}
\description{
Get the top albums listened to by a user. You can stipulate a time period. Sends the overall chart by default.
}
\seealso{
Other raw user functions: 
\code{\link{raw_user_get_friends}()},
\code{\link{raw_user_get_info}()},
\code{\link{raw_user_get_loved_tracks}()},
\code{\link{raw_user_get_personal_tags}()},
\code{\link{raw_user_get_recent_tracks}()},
\code{\link{raw_user_get_top_artists}()},
\code{\link{raw_user_get_top_tags}()},
\code{\link{raw_user_get_top_tracks}()}
}
\concept{raw user functions}
