% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{raw_user_get_top_tracks}
\alias{raw_user_get_top_tracks}
\title{Get the top tracks listened to by a user. You can stipulate a time period.
Sends the overall chart by default.}
\usage{
raw_user_get_top_tracks(
  username,
  period = "7day",
  .start_page = 1,
  .n_pages = NULL
)
}
\arguments{
\item{username}{The user name to fetch top tracks for.}

\item{period}{overall | 7day | 1month | 3month | 6month | 12month - The time
period over which to retrieve top tracks for.}

\item{.start_page}{the page to start pulling from, by default starts from the
1st page}

\item{.n_pages}{the total number of pages to pull. If no value is provided,
all pages are retrieved.}
}
\value{
A LastFM API object
}
\description{
Get the top tracks listened to by a user. You can stipulate a time period.
Sends the overall chart by default.
}
\seealso{
Other raw user functions: 
\code{\link{raw_user_get_friends}()},
\code{\link{raw_user_get_info}()},
\code{\link{raw_user_get_loved_tracks}()},
\code{\link{raw_user_get_personal_tags}()},
\code{\link{raw_user_get_recent_tracks}()},
\code{\link{raw_user_get_top_albums}()},
\code{\link{raw_user_get_top_artists}()},
\code{\link{raw_user_get_top_tags}()}
}
\concept{raw user functions}
