% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastfm.R
\name{single-page-endpoint}
\alias{single-page-endpoint}
\title{Last.FM querying endpoint}
\arguments{
\item{method}{a string}

\item{result_node}{a string}

\item{.page}{Page to query, by default query the first page}
}
\value{
A lastfmr API object, see \code{\link[=lastfmr]{lastfmr()}} for information.
}
\description{
Used for documenting single-query endpoints
}
