% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track_getSimilar}
\alias{track_getSimilar}
\title{Get the similar tracks for this track.}
\usage{
track_getSimilar(track, artist, mbid = NA, autocorrect = NA, limit = NA,
  api_key = lastkey)
}
\arguments{
\item{track}{The track name.}

\item{artist}{The artist name.}

\item{mbid}{The musicbrainz id for the track}

\item{autocorrect}{Transform misspelled artist and track names into correct
artist and track names, returning the correct version instead.
The corrected artist and track name will be returned in the response. [0|1]}

\item{limit}{Maximum number of similar tracks to return.}

\item{api_key}{A Last.fm API key.}
}
\value{
A list of the similar tracks.
}
\description{
Get the similar tracks for this track on Last.fm, based on listening data.
Implementation of last.fm's \emph{track.getSimilar} API method
(\url{http://www.last.fm/api/show/track.getSimilar})
}
\examples{
track_getSimilar("Hard Times", "Thundercat")
}
