% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user-methods}
\alias{user-methods}
\alias{raw_user_get_recent_tracks}
\alias{user_get_recent_tracks}
\alias{raw_user_get_top_albums}
\alias{user_get_top_albums}
\alias{raw_user_get_top_artists}
\alias{user_get_top_artists}
\alias{raw_user_get_top_tracks}
\alias{user_get_top_tracks}
\alias{raw_user_get_friends}
\alias{raw_user_get_info}
\alias{raw_user_get_loved_tracks}
\alias{raw_user_get_personal_tags}
\alias{raw_user_get_top_tags}
\title{User Methods}
\usage{
raw_user_get_recent_tracks(
  username,
  from = NULL,
  to = NULL,
  extended = 0,
  .start_page = 1,
  .n_pages = NULL
)

user_get_recent_tracks(
  username,
  from = NULL,
  to = NULL,
  extended = 0,
  .start_page = 1,
  .n_pages = NULL
)

raw_user_get_top_albums(
  username,
  period = "7day",
  .start_page = 1,
  .n_pages = NULL
)

user_get_top_albums(
  username,
  period = "7day",
  .start_page = 1,
  .n_pages = NULL
)

raw_user_get_top_artists(
  username,
  period = "7day",
  .start_page = 1,
  .n_pages = NULL
)

user_get_top_artists(
  username,
  period = "7day",
  .start_page = 1,
  .n_pages = NULL
)

raw_user_get_top_tracks(
  username,
  period = "7day",
  .start_page = 1,
  .n_pages = NULL
)

user_get_top_tracks(
  username,
  period = "7day",
  .start_page = 1,
  .n_pages = NULL
)

raw_user_get_friends(
  username,
  recent_tracks = TRUE,
  .start_page = 1,
  .n_pages = NULL
)

raw_user_get_info(username = NULL, .page = 1)

raw_user_get_loved_tracks(username, .start_page = 1, .n_pages = NULL)

raw_user_get_personal_tags(
  username,
  tag,
  tagging_type,
  .start_page = 1,
  .n_pages = NULL
)

raw_user_get_top_tags(username, .page = 1)
}
\arguments{
\item{username}{The username as a string}

\item{from}{(Optional) Beginning timestamp of a range - only display
scrobbles after this time.}

\item{to}{(Optional) End timestamp of a range - only display scrobbles before
this time.}

\item{extended}{(Optional) Whether to include extended data in each artist, and whether
or not the user has loved each track.}

\item{.start_page}{The page to start pulling from, by default starts from the
1st page}

\item{.n_pages}{The total number of pages to pull. If no value is provided,
all pages are retrieved.}

\item{period}{(Optional) The time period to aggregate tracks/albums/artists over as a string.
One of overall, 7day, 1month, 3month, 6month, 12month. \verb{7day} by default.}

\item{recent_tracks}{(Optional) Whether to include recent tracks fro friends in the response}

\item{tag}{The tag you're interested in.}

\item{tagging_type}{The objects on which the tags were applied as a string}
}
\description{
These functions retrieve information about Last.fm users. Users are identified
by \code{username}:
\itemize{
\item \code{\link[=raw_user_get_recent_tracks]{raw_user_get_recent_tracks()}} retrieve a list of recent tracks scrobbled by the user.
\item \code{\link[=raw_user_get_top_albums]{raw_user_get_top_albums()}}, \code{\link[=raw_user_get_top_artists]{raw_user_get_top_artists()}},
and \code{\link[=raw_user_get_top_tracks]{raw_user_get_top_tracks()}} retrieve lists of top albums, artists,
and tracks respectively played by the user.
}
}
\concept{raw user functions}
\concept{tidy user functions}
