% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user_get_top_tracks}
\alias{user_get_top_tracks}
\title{Title}
\usage{
user_get_top_tracks(
  username,
  period = "7day",
  .start_page = 1,
  .n_pages = NULL
)
}
\arguments{
\item{username}{}

\item{period}{}

\item{.start_page}{the page to start pulling from, by default starts from the
1st page}

\item{.n_pages}{the total number of pages to pull. If no value is provided,
all pages are retrieved.}
}
\value{

}
\description{
Title
}
