% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user_getWeeklyAlbumChart}
\alias{user_getWeeklyAlbumChart}
\title{Get an album chart for a user profile.}
\usage{
user_getWeeklyAlbumChart(user, from = NA, to = NA)
}
\arguments{
\item{user}{The last.fm username to fetch the charts of.}

\item{from}{The date at which the chart should start from.}

\item{to}{The date at which the chart should end on.}
}
\value{
A list of an album chart for a user profile.
}
\description{
Get an album chart for a user profile, for a given date range.
If no date range is supplied, it will return the most recent album chart for this user.
Implementation of last.fm's \emph{user.getWeeklyAlbumChart} API method
(\url{http://www.last.fm/api/show/user.getWeeklyAlbumChart}).
}
\examples{
\dontrun{
user_getWeeklyAlbumChart("platyjus")
}
}
