% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{userGetArtistTracks}
\alias{userGetArtistTracks}
\title{Title}
\usage{
userGetArtistTracks(artist, api_key, user, limit = NULL, page = NULL)
}
\arguments{
\item{artist}{(character; Required) : The artist to search for.}

\item{api_key}{(apiAccount; Required) : A Last.fm API key.}

\item{user}{(apiUser; Required) : The user name to fetch top albums for.}

\item{limit}{(int; Optional) : The number of results to fetch per page.
Defaults to 50.}

\item{page}{(int; Optional) : The page number to fetch. Defaults to first
page.}
}
\value{
A \code{\link[tibble]{tibble}} of results
}
\description{
Title
}
